Extract profile utilities module
Move SPIFFE URI and profile helper functions into profile.rs with a
shared format constant. Update daemon/main callers and tests to use the
new module.

Closes #35
